+++ test/Virtlink.Utilib.Tests/IO/DataBufferTests.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Virtlink.Utilib.IO
+{
+    /// <summary>
+    /// Tests the <see cref="DataBuffer"/> class.
+    /// </summary>
+    public sealed partial class DataBufferTests
+    {
+    }
+}