+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Add `IBuffer` interface and `DataBuffer` class, which represents a buffer of bytes.
+
 ## [1.23.0] - 2017-05-23
 - Add `SetComparer` for comparing sets.